Added end of round and end game logic
At the end of the round, a new deck is built, shuffled, burned, and
dealt. Game statuses such as Handmaid and stillInRound are reset to
normal.

Game will display leader board and end the game when a player has won.

WIP Edge case for prince card when a player needs to draw a card and the
deck is empty